requirements.txt
requests==2.*
pyaml

